complete fundamentals
