+++ fundamentals/src/books.js
@@ -1,0 +1,14 @@
+export const books = [
+  {
+    title: "Fourth Wing (The Empyrean, 1)",
+    author: "Rebecca Yarros",
+    img: "https://images-na.ssl-images-amazon.com/images/I/91n7p-j5aqL._AC_UL600_SR600,400_.jpg",
+    id: 1,
+  },
+  {
+    title: "The Woman in Me",
+    author: "Britney Spears ",
+    img: "https://images-na.ssl-images-amazon.com/images/I/61BWsc9eGbL._AC_UL600_SR600,400_.jpg",
+    id: 2,
+  },
+];
