+++ fundamentals/src/index.css
@@ -30,6 +30,7 @@ body {
   border-radius: 1rem;
   padding: 2rem;
   text-align: center;
+  position: relative;
 }
 .book img {
   width: 100%;
@@ -49,8 +50,20 @@ body {
 .book p {
   margin: 1rem 0 0.5rem;
 }
+
 h1 {
   text-align: center;
   margin-top: 4rem;
   text-transform: capitalize;
+}
+
+.number {
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding: 0.75rem;
+  font-size: 1rem;
+  border-top-left-radius: 1rem;
+  border-bottom-right-radius: 1rem;
+  background-color: #c35600;
 }
